Add a dedicated index page
This is nicer than just defaulting to the outside page and more
useful, especially for manually exploring what's on the server.

Also add a test script which checks the setup of the machine based
on the links found on the page. This is handy to check that things
are up and running.

Change-Id: Iaff29f867635916c6a59da17b350dff740e30ac6