Update the system package list before installs
This seems to make the installs more reliable, though I'm not really
sure what causes the issue I've seen without it (specifically that
some of the package downloads fail). I'd guess that the locations
of the download servers has changed and that something's caching
their old locations in the base box; this is however speculation.

Change-Id: I746462be06e04c6f3971d158a1910e577a1d7ba0