Make the generate_screens.py script run agin after failure
Previously the 'refreshonly' attribute meant that this command could
fail to run even after it had failed the first time. Introduce a
file creation after the runs to validate that it was successful.
Also remove the file before each run so that re-runs (which we do
want to happen) are suitably picked up too.

Change-Id: I7740b015bbf21556b7f31fb8fe6b0ff5f9376577