Install Sphinx from the repos
Sphinx and its dependencies take a considerable amount of effort
to install from PyPI, but are only actually needed for building
the docs (and thus part of the 'setup_requires' rather than
'install_requires' of the srcomp suite. Since that effort is
undesirable on a Raspberry Pi and there's a compatible version in
the Debian repos, we just use that instead.