Make the update command work under Python 3
This changes the approach used to one which should be more portable.
We now make use of the installed srcomp-http package, whose 'update'
module we just run as a module. This turns out to already have been
supported, meaning we don't even need the local clone of the repo.

We also change from passing the path to a "virtual environment",
which hasn't made sense in this configuration for a long time,
to just passing the full path to the Python executable to use.
This provides more control to the manifest and we use that to
change to using the system Python 3 (rather than the Python 2 we
were previously using).