Add support for secondary TLS domain names
This adds support for creating a TLS certificate with more than
one domain on it. This allows servers to operate securely at the
end of one of a number of domains.

Nginx remains unaware of the other domains, however since it was
already happily serving to all of them this isn't a change.

Fixes https://github.com/PeterJCLaw/srcomp-puppet/issues/23